[MDS-5410] Minespace Management Button (#2622)
* update nvmrc with new node version, and for restart codespaces add a yarn install and nvm install. Fix is in DashboardRoutes to take out accidentally added 'exact' attribute. Made some changes to deal with console errors: switching AdminNavigation to use items instead of children got rid of that warning, and also react-generated errors about warnKey and eventKey. CSS updates to make it the same as it was before, as the change to items resulted in different classname structure

* update snaps
