+++ .nvmrc
@@ -1,1 +1,1 @@
-v14.8.0
+v14.21.3
