+++ services/core-web/src/styles/index.scss
@@ -46,5 +46,6 @@
 @import "./components/NotificationDrawer.scss";
 @import "./components/SteppedForm.scss";
 @import "./components/DocumentTableWithExpandedRows.scss";
+@import "./components/Menu.scss";
 
 // UTILITIES - utilities and helper classes. This layer has the highest specificity.
